Complete lesson flex-layout-2
I'm learning that the best way to target an element is in the most
direct way possible. This lesson is a testament to that, as while all
the things I did got me to the same place, some of the ways I went
about getting there were far more convoluted then they needed to be.